Issue #17
Make getUrl smarter about absolute and relative urls in the request
line. The current implementation has a tendancy to prepend the hostname
to absolute urls, which leads to unusable data in the AWS UI. The new
version will not do concatenation if the url in the request is already
absolute.